requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
google-auth==2.49.1
google-auth-oauthlib==1.3.1
gspread==6.2.1
idna==3.11
oauthlib==3.3.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pypdfium2==5.6.0
pyTelegramBotAPI==4.32.0
python-dotenv==1.2.2
PySocks==1.7.1
requests==2.33.1
requests-oauthlib==2.0.0
urllib3==2.6.3
